+++ zcode/inout/__init__.py
@@ -7,8 +7,11 @@ from . import log
 from .log import *
 from . import timer
 from .timer import *
+from . import singleton
+from .singleton import *
 
 __all__ = []
 __all__.extend(inout_core.__all__)
 __all__.extend(log.__all__)
 __all__.extend(timer.__all__)
+__all__.extend(singleton.__all__)